# src/gv_server/models/game.py
# coding: utf-8

"""
    Game Voting API

    This API is meant as baseline functionality for a game voting tool that is combined with an Asyncapi specification for all the interative pieces.

    The version of the OpenAPI document: 2.0.0
    Contact: philipp.schmurr@gmail.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Game(BaseModel):
    """
    Game
    """ # noqa: E501
    id: StrictStr = Field(alias="ID")
    name: StrictStr = Field(alias="Name")
    cover: StrictStr = Field(alias="Cover")
    genre: List[StrictStr] = Field(alias="Genre")
    mp: StrictStr = Field(alias="MP")
    type: StrictStr = Field(alias="Type")
    toplevel: StrictStr = Field(alias="Toplevel")
    meta: List[StrictStr] = Field(alias="Meta")
    parent: Optional[StrictStr] = Field(alias="Parent")
    children: List[StrictStr] = Field(alias="Children")
    steam_appid: Optional[StrictStr] = Field(default=None, alias="SteamAppid")
    detailed_description: Optional[StrictStr] = None
    description: Optional[StrictStr] = None
    categories: Optional[List[StrictStr]] = None
    genres: Optional[List[StrictStr]] = None
    release_date: Optional[datetime] = None
    readme: Optional[StrictStr] = Field(default=None, alias="Readme")
    __properties: ClassVar[List[str]] = ["ID", "Name", "Cover", "Genre", "MP", "Type", "Toplevel", "Meta", "Parent", "Children", "SteamAppid", "detailed_description", "description", "categories", "genres", "release_date", "Readme"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Game from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # set to None if parent (nullable) is None
        # and model_fields_set contains the field
        if self.parent is None and "parent" in self.model_fields_set:
            _dict['Parent'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Game from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        # raise errors for additional fields in the input
        for _key in obj.keys():
            if _key not in cls.__properties:
                raise ValueError("Error due to additional fields (not defined in Game) in the input: " + _key)

        _obj = cls.model_validate({
            "ID": obj.get("ID"),
            "Name": obj.get("Name"),
            "Cover": obj.get("Cover"),
            "Genre": obj.get("Genre"),
            "MP": obj.get("MP"),
            "Type": obj.get("Type"),
            "Toplevel": obj.get("Toplevel"),
            "Meta": obj.get("Meta"),
            "Parent": obj.get("Parent"),
            "Children": obj.get("Children"),
            "SteamAppid": obj.get("SteamAppid"),
            "detailed_description": obj.get("detailed_description"),
            "description": obj.get("description"),
            "categories": obj.get("categories"),
            "genres": obj.get("genres"),
            "release_date": obj.get("release_date"),
            "Readme": obj.get("Readme")
        })
        return _obj