# src/gv_server/models/game_votes.py
# coding: utf-8

"""
    Game Voting API

    This API is meant as baseline functionality for a game voting tool that is combined with an Asyncapi specification for all the interative pieces.

    The version of the OpenAPI document: 2.0.0
    Contact: philipp.schmurr@gmail.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class GameVotes(BaseModel):
    """
    GameVotes
    """ # noqa: E501
    game_id: StrictStr
    likes: StrictInt
    abstains: StrictInt
    dislikes: StrictInt
    __properties: ClassVar[List[str]] = ["game_id", "likes", "abstains", "dislikes"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of GameVotes from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of GameVotes from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        # raise errors for additional fields in the input
        for _key in obj.keys():
            if _key not in cls.__properties:
                raise ValueError("Error due to additional fields (not defined in GameVotes) in the input: " + _key)

        _obj = cls.model_validate({
            "game_id": obj.get("game_id"),
            "likes": obj.get("likes"),
            "abstains": obj.get("abstains"),
            "dislikes": obj.get("dislikes")
        })
        return _obj